let user set parallel test timeout

+++ pyuvsim/tests/conftest.py
@@ -44,7 +44,8 @@ def pytest_configure(config):
     """Register an additional marker."""
     config.addinivalue_line(
         "markers",
-        "parallel(n): mark test to run in n parallel mpi processes."
+        "parallel(n, timeout=70): mark test to run in n parallel mpi processes."
+        " Optionally, set a timeout in seconds."
     )
 
 
@@ -82,11 +83,15 @@ def pytest_runtest_call(item):
     if parmark is None:
         return
     nproc = 1
-    if len(parmark.args) == 1:
+    if len(parmark.args) >= 1:
         try:
             nproc = int(parmark.args[0])
         except ValueError:
             raise ValueError(f"Invalid number of processes: {parmark.args[0]}")
+
+    timeout = 70
+    if 'timeout' in parmark.kwargs:
+        timeout = float(parmark.kwargs['timeout'])
 
     call = ['mpirun', '--host', 'localhost:10', '-n', str(nproc),
             "python", "-m", "pytest", "{:s}::{:s}".format(str(item.fspath), str(item.name))]
@@ -94,11 +99,14 @@ def pytest_runtest_call(item):
         try:
             envcopy = os.environ.copy()
             envcopy['TEST_IN_PARALLEL'] = '1'
-            check_output(call, env=envcopy, stderr=DEVNULL, timeout=70)
+            check_output(call, env=envcopy, stderr=DEVNULL, timeout=timeout)
         except (TimeoutExpired, CalledProcessError) as err:
             message = f"Parallelized test {item.name} failed"
             if isinstance(err, TimeoutExpired):
-                message += " due to timeout"
+                message += (f" after {timeout} seconds due to timeout. \nThe timeout may be set"
+                            " via the ``timeout`` keyword in the ``parallel`` decorator. \n"
+                            "A stalled test may be caused by an inconsistent MPI state. Check"
+                            " that blocking operations are reached by all processes")
             message += "."
             raise AssertionError(message) from err
 
@@ -142,18 +150,19 @@ def pytest_runtest_makereport(item, call):
     # Is a parallel test but not currently within the subprocess.
     if report.failed and (not issubproc) and (parmark is not None) and (report.when == 'call'):
         pth = f"/tmp/mpitest_{report.head_line}"
-        for ii, repf in enumerate(os.listdir(pth)):
-            if not repf.endswith(".pkl"):
-                continue
-            rank = int(re.findall('(?<=rank)[0-9]+', repf)[0])
-            rank_report = pkl.load(open(os.path.join(pth, repf), 'rb'))
-            os.remove(os.path.join(pth, repf))
-            if ii == 0:
-                report = rank_report
-                report.longrepr = _MPIExceptionChainRepr(report.longrepr, rank)
-            else:
-                report.longrepr.append(rank_report.longrepr, rank)
-        os.rmdir(pth)
+        if os.path.exists(pth):
+            for ii, repf in enumerate(os.listdir(pth)):
+                if not repf.endswith(".pkl"):
+                    continue
+                rank = int(re.findall('(?<=rank)[0-9]+', repf)[0])
+                rank_report = pkl.load(open(os.path.join(pth, repf), 'rb'))
+                os.remove(os.path.join(pth, repf))
+                if ii == 0:
+                    report = rank_report
+                    report.longrepr = _MPIExceptionChainRepr(report.longrepr, rank)
+                else:
+                    report.longrepr.append(rank_report.longrepr, rank)
+            os.rmdir(pth)
         nproc = int(parmark.args[0])
         outcome = call.excinfo.exconly()
         mpitest_info = f"Run with {nproc} processes."
